# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API untuk sistem manajemen rental PlayStation Byone Arena. Mendukung realtime via WebSocket untuk aplikasi mobile dan TV Android.",
        "title": "Byone Arena API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Byone Arena Support",
            "email": "support@byone-arena.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/vouchers": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengembalikan daftar seluruh voucher diskon beserta status dan statistik penggunaan",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Voucher"
                ],
                "summary": "Ambil semua voucher",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/entity.Voucher"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Hanya admin dan superadmin",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Membuat voucher diskon baru.\\n\\n**Nilai discountType yang valid:**\\n- `percentage` — diskon persen dari total (discountValue = 0–100)\\n- `fixed_amount` — diskon nominal Rp tetap\\n\\n**Format expiresAt:** RFC3339/ISO8601, contoh `2026-06-26T00:00:00Z`. Kirim `null` jika tidak ada batas waktu.\\n\\nKode otomatis diubah ke UPPERCASE.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Voucher"
                ],
                "summary": "Buat voucher baru",
                "parameters": [
                    {
                        "description": "Data voucher baru",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/usecase.CreateVoucherRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/entity.Voucher"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Validasi gagal — cek discountType (percentage|fixed_amount) dan format expiresAt (RFC3339)",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Hanya admin atau superadmin",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/vouchers/code/{code}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengecek validitas dan detail voucher berdasarkan kodenya. Dapat digunakan kasir untuk preview diskon sebelum pembayaran.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Voucher"
                ],
                "summary": "Cek voucher berdasarkan kode",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Kode voucher (tidak case-sensitive)",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/entity.Voucher"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Kode tidak ditemukan",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/vouchers/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengembalikan detail satu voucher diskon",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Voucher"
                ],
                "summary": "Ambil voucher berdasarkan ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Voucher ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/entity.Voucher"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Memperbarui informasi voucher. Semua field bersifat opsional.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Voucher"
                ],
                "summary": "Update data voucher",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Voucher ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Data voucher yang diperbarui",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/usecase.UpdateVoucherRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/entity.Voucher"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Menghapus voucher secara permanen. Voucher yang sudah dipakai di transaksi tidak dapat dihapus.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Voucher"
                ],
                "summary": "Hapus voucher",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Voucher ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/vouchers/{id}/toggle": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Toggle status isActive voucher. Voucher nonaktif tidak bisa digunakan saat pembayaran.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Voucher"
                ],
                "summary": "Aktifkan / nonaktifkan voucher",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Voucher ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/entity.Voucher"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "entity.DiscountType": {
            "type": "string",
            "enum": [
                "percentage",
                "fixed_amount"
            ],
            "x-enum-comments": {
                "DiscountTypeFixedAmount": "diskon nominal tetap",
                "DiscountTypePercentage": "diskon persen dari total"
            },
            "x-enum-varnames": [
                "DiscountTypePercentage",
                "DiscountTypeFixedAmount"
            ]
        },
...
        "entity.Voucher": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "discountType": {
                    "$ref": "#/definitions/entity.DiscountType"
                },
                "discountValue": {
                    "type": "number"
                },
                "expiresAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isActive": {
                    "type": "boolean"
                },
                "maxDiscount": {
                    "description": "batas maks diskon persen (0 = tidak terbatas)",
                    "type": "number"
                },
                "maxUsage": {
                    "description": "batas total pemakaian (0 = tidak terbatas)",
                    "type": "integer"
                },
                "minPurchase": {
                    "description": "minimal total sebelum voucher berlaku",
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "usageCount": {
                    "type": "integer"
                }
            }
        },
        "response.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "response.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "usecase.CreateVoucherRequest": {
            "type": "object",
            "required": [
                "code",
                "discountType",
                "discountValue",
                "name"
            ],
            "properties": {
                "code": {
                    "description": "Kode unik voucher (otomatis uppercase). Contoh: DISKON10",
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 3,
                    "example": "HPH3"
                },
                "discountType": {
                    "description": "Tipe diskon: \"percentage\" (persen) atau \"fixed_amount\" (nominal Rp)",
                    "enum": [
                        "percentage",
                        "fixed_amount"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/entity.DiscountType"
                        }
                    ],
                    "example": "fixed_amount"
                },
                "discountValue": {
                    "description": "Nilai diskon: jika percentage maka 0-100, jika fixed_amount maka nominal Rp",
                    "type": "number",
                    "example": 24000
                },
                "expiresAt": {
                    "description": "Tanggal kedaluwarsa voucher dalam format RFC3339/ISO8601. Contoh: 2026-06-26T00:00:00Z (null = tidak ada batas waktu)",
                    "type": "string",
                    "example": "2026-06-26T00:00:00Z"
                },
                "maxDiscount": {
                    "description": "Batas maksimal nominal diskon untuk tipe percentage (0 = tidak terbatas)",
                    "type": "number",
                    "minimum": 0,
                    "example": 0
                },
                "maxUsage": {
                    "description": "Batas maksimal total pemakaian voucher (0 = tidak terbatas)",
                    "type": "integer",
                    "minimum": 0,
                    "example": 100
                },
                "minPurchase": {
                    "description": "Minimal total belanja sebelum voucher berlaku (0 = tidak ada minimum)",
                    "type": "number",
                    "minimum": 0,
                    "example": 24000
                },
                "name": {
                    "description": "Nama deskriptif voucher",
                    "type": "string",
                    "maxLength": 150,
                    "minLength": 3,
                    "example": "Happy Hours 3 Jam"
                }
            }
        },
...
        "usecase.UpdateVoucherRequest": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 3,
                    "example": "HPH3"
                },
                "discountType": {
                    "description": "Tipe diskon: \"percentage\" atau \"fixed_amount\"",
                    "enum": [
                        "percentage",
                        "fixed_amount"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/entity.DiscountType"
                        }
                    ],
                    "example": "fixed_amount"
                },
                "discountValue": {
                    "type": "number",
                    "example": 24000
                },
                "expiresAt": {
                    "description": "Format RFC3339/ISO8601. Contoh: 2026-06-26T00:00:00Z",
                    "type": "string",
                    "example": "2026-06-26T00:00:00Z"
                },
                "isActive": {
                    "type": "boolean"
                },
                "maxDiscount": {
                    "type": "number",
                    "minimum": 0
                },
                "maxUsage": {
                    "type": "integer",
                    "minimum": 0
                },
                "minPurchase": {
                    "type": "number",
                    "minimum": 0
                },
                "name": {
                    "type": "string",
                    "maxLength": 150,
                    "minLength": 3,
                    "example": "Happy Hours 3 Jam"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Masukkan token JWT dengan format: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}